Implement MPC Planner, Optimizer, and Polyfitter classes
- Added MPCPlanner class for planning vehicle control commands based on current state and waypoints.
- Introduced MPCOptimizer class to solve the optimal control problem using Model Predictive Control (MPC) techniques.
- Developed Polyfitter class for lane detection and fitting, including methods for image processing, clustering, and polynomial fitting.
- Implemented methods for converting image points to world coordinates and preparing reference waypoints for MPC.
- Enhanced lane detection capabilities with clustering and curve fitting algorithms.
- Added utility functions for image loading, displaying results, and computing virtual centerlines.

+++ car_controls/includes/inference/LaneCurveFitter.hpp
@@ -17,7 +17,7 @@ public:
 		std::vector<cv::Point2f> c2;
 	};
 
-	LaneCurveFitter(float dbscanEps = 5.0f, int dbscanMinSamples = 20, int numWindows = 20, int laneWidthPx = 300);
+	LaneCurveFitter(float dbscanEps = 5.0f, size_t dbscanMinSamples = 20, int numWindows = 20, int laneWidthPx = 300);
 
 	std::vector<LaneCurve> fitLanes(const cv::Mat& binaryMask);
 	std::optional<CenterlineResult> computeVirtualCenterline(const std::vector<LaneCurve>& lanes, int imgWidth, int imgHeight);
